requirements.txt
streamlit
python-dotenv
groq
streamlit-mic-recorder
pydub
openai
mcp
asyncio
uuid